# back/docs/swagger.yaml
basePath: /
definitions:
...
  model.Diagram:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      modelId:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      workspaceId:
        type: string
    required:
    - name
    type: object
...
  model.Workspace:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      diagrams:
        items:
          $ref: '#/definitions/model.Diagram'
        type: array
      id:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  request.WorkspaceRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for Easy DEVS.
  title: Easy DEVS API
  version: "1.0"
paths:
...
  /workspace:
    get:
      description: Retrieve a list of all workspace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Workspace'
            type: array
      summary: Get all workspace
      tags:
      - workspace
    post:
      consumes:
      - application/json
      description: Create a new workspace and store it in the database
      parameters:
      - description: Workspace object
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/request.WorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Workspace'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new workspace
      tags:
      - workspace
  /workspace/{id}:
    delete:
      description: Delete an existing workspace by its ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a workspace by ID
      tags:
      - workspace
    get:
      description: Retrieve a single workspace by its ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Workspace'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a workspace by ID
      tags:
      - workspace
    patch:
      consumes:
      - application/json
      description: Update an existing workspace with partial data
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Partial workspace update
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/request.WorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Workspace'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update a workspace
      tags:
      - workspace
swagger: "2.0"